Add modal & confirm modal, finish delete in admin

+++ backend/controllers/ratingController.js
@@ -1,4 +1,5 @@
 const Rating = require("../models/ratingModel");
+const User = require("../models/userModel");
 const mongoose = require("mongoose");
 
 exports.getAllRatings = async (req, res) => {
@@ -80,11 +81,19 @@ exports.updateRating = async (req, res) => {
 exports.deleteRating = async (req, res) => {
     const { id } = req.params
     if (!mongoose.Types.ObjectId.isValid(id)) return res.status(404).json({error: 'No rating found for id ' + id});
+
+    if (!req.user || !req.user.id)
+        return res.status(404).json({error: 'User not logged in'});
+
+    const user = await User.findById(req.user.id);
+    if (!user) 
+        return res.status(401).json({error: 'User not found'});
+
+    const ratingCheck = await Rating.findById(id);
+    if (!ratingCheck) return res.status(404).json({error: 'No rating found for id ' + id});
     
-    const rating = await Rating.findOneAndDelete({_id: id});
-    if(!rating) return res.status(404).json({error: 'No rating found for id ' + id});
-   
-    res.status(200).json({message: 'Rating deleted.'});
+    await Rating.findOneAndDelete({_id: id});
+    res.status(200).json({ success: true, deletedId: id });
 }
 
 const fetchRatingsWithUsers = async (filter) => {
